feat: Search IntegrationTest 분리
