+++ src/main/kotlin/com/flab/ticketing/performance/dto/service/PerformanceSummarySearchResult.kt
@@ -1,5 +1,6 @@
 package com.flab.ticketing.performance.dto.service
 
+import com.flab.ticketing.performance.entity.Performance
 import java.time.LocalDate
 import java.time.ZonedDateTime
 
@@ -21,5 +22,20 @@ data class PerformanceSummarySearchResult(
     ) : this(
         uid, image, title, regionName, startDate?.toLocalDate(), endDate?.toLocalDate()
     )
+
+    companion object {
+
+        fun of(performance: Performance): PerformanceSummarySearchResult {
+            return PerformanceSummarySearchResult(
+                performance.uid,
+                performance.image,
+                performance.name,
+                performance.regionName,
+                performance.performanceDateTime.minOf { it.showTime },
+                performance.performanceDateTime.maxOf { it.showTime },
+            )
+
+        }
+    }
 
 }
